outsourced downloading of swbFiles and adapted logic of agb component
