+++ anzeigen-frontend/src/components/Ad/Details/AdOverview.vue
@@ -158,8 +158,11 @@ import AdDisplayCard from "@/components/common/AdDisplayCard.vue";
 import AdDisplaySheet from "@/components/common/AdDisplaySheet.vue";
 import IconText from "@/components/common/IconText.vue";
 import { useGetFile } from "@/composables/api/useFilesApi";
+import { useDownloadFile } from "@/composables/useDownloadFile.ts";
 import { DATE_DISPLAY_FORMAT } from "@/Constants";
 import router from "@/plugins/router";
+
+const downloadFile = useDownloadFile();
 
 const { adDetails } = defineProps<{
   adDetails: Readonly<AdTO>;
@@ -189,25 +192,25 @@ const routeToUser = (id: number) => {
   });
 };
 
-/**
- * Downloads a file based on the provided ID.
- * Retrieves the file, creates a Blob, and triggers a download.
- * @param id - The ID of the file to download.
- */
-const downloadFile = async (id: number) => {
-  await getFile({ id: id });
-
-  if (fileData.value && fileData.value.fileBase64 && fileData.value.name) {
-    const blob = new Blob([fileData.value?.fileBase64]);
-    const fileURL = URL.createObjectURL(blob);
-    const downloadLink = document.createElement("a");
-
-    downloadLink.href = fileURL;
-    downloadLink.download = fileData.value.name;
-    document.body.appendChild(downloadLink);
-    downloadLink.click();
-  }
-};
+///**
+// * Downloads a file based on the provided ID.
+// * Retrieves the file, creates a Blob, and triggers a download.
+// * @param id - The ID of the file to download.
+// */
+//const downloadFile = async (id: number) => {
+//  await getFile({ id: id });
+//
+//  if (fileData.value && fileData.value.fileBase64 && fileData.value.name) {
+//    const blob = new Blob([fileData.value?.fileBase64]);
+//    const fileURL = URL.createObjectURL(blob);
+//    const downloadLink = document.createElement("a");
+//
+//    downloadLink.href = fileURL;
+//    downloadLink.download = fileData.value.name;
+//    document.body.appendChild(downloadLink);
+//    downloadLink.click();
+//  }
+//};
 </script>
 
 <style scoped>
